# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for F1 Soluções Financeiras
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SubmitLeadBody {
  /** @minLength 2 */
  name: string;
  email: string;
  phone: string;
  /** @nullable */
  cpf?: string | null;
  /** @nullable */
  licensePlate?: string | null;
  /** @nullable */
  vehiclePaid?: boolean | null;
  /** @nullable */
  vehicleInOwnerName?: boolean | null;
  /** @nullable */
  hasIncome?: boolean | null;
  consentLgpd: boolean;
  /** @nullable */
  source?: string | null;
}

export interface LeadResponse {
  id: number;
  message: string;
}

export interface LeadsStats {
  total: number;
  today: number;
  thisWeek: number;
  thisMonth: number;
}

export interface SubmitContactBody {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  subject: string;
  message: string;
  consentLgpd: boolean;
}

export interface ContactResponse {
  id: number;
  message: string;
}

export interface ErrorResponse {
  error: string;
}