# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for F1 Soluções Financeiras
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Capture a potential customer lead from the website form
 * @summary Submit a new lead
 */
export const submitLeadBodyNameMin = 2;

export const SubmitLeadBody = zod.object({
  name: zod.string().min(submitLeadBodyNameMin),
  email: zod.string().email(),
  phone: zod.string(),
  cpf: zod.string().nullish(),
  licensePlate: zod.string().nullish(),
  vehiclePaid: zod.boolean().nullish(),
  vehicleInOwnerName: zod.boolean().nullish(),
  hasIncome: zod.boolean().nullish(),
  consentLgpd: zod.boolean(),
  source: zod.string().nullish(),
});

/**
 * Returns aggregated lead statistics
 * @summary Get lead statistics
 */
export const GetLeadsStatsResponse = zod.object({
  total: zod.number(),
  today: zod.number(),
  thisWeek: zod.number(),
  thisMonth: zod.number(),
});

/**
 * Submit a contact/inquiry form message
 * @summary Submit a contact message
 */
export const submitContactBodyNameMin = 2;

export const SubmitContactBody = zod.object({
  name: zod.string().min(submitContactBodyNameMin),
  email: zod.string().email(),
  phone: zod.string().nullish(),
  subject: zod.string(),
  message: zod.string(),
  consentLgpd: zod.boolean(),
});